requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
comtypes==1.4.9
distro==1.9.0
graphviz==0.20.3
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
keyboard==0.13.5
MouseInfo==0.1.3
ollama==0.4.7
pillow==11.1.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
pyttsx3==2.98
pytweening==1.2.0
pywin32==308
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.1
typing_extensions==4.12.2
